refactor - organize types/props
