+++ src/components/screens/paidMemberMng/matterForm/create/FirstForm/index.tsx
@@ -159,9 +159,7 @@ export const Contents: VFC<ContentsType> = ({ onSubmit, onError }) => {
   );
 };
 
-export const FirstForm: VFC<{ onSubmit: (data: FormData) => void }> = ({
-  onSubmit,
-}) => {
+export const FirstForm: VFC<Omit<ContentsType, "onError">> = ({ onSubmit }) => {
   const onError = (errors: any) => {
     console.log(`execute: onError`);
     console.log(errors);
