+++ src/components/screens/paidMemberMng/matterForm/create/OnboardingForm/index.tsx
@@ -29,9 +29,9 @@ export const Contents: VFC<ContentsType> = ({ onSubmit, onError }) => {
   );
 };
 
-export const OnboardingForm: VFC<{
-  onSubmit: (data: FormData) => void;
-}> = ({ onSubmit }) => {
+export const OnboardingForm: VFC<Omit<ContentsType, "onError">> = ({
+  onSubmit,
+}) => {
   const onError = (errors: any) => {
     console.log(`execute: onError`);
     console.log(errors);
