+++ src/components/screens/paidMemberMng/matterForm/create/PaymentMethodForm/index.tsx
@@ -29,9 +29,9 @@ export const Contents: VFC<ContentsType> = ({ onSubmit, onError }) => {
   );
 };
 
-export const PaymentMethodForm: VFC<{
-  onSubmit: (data: FormData) => void;
-}> = ({ onSubmit }) => {
+export const PaymentMethodForm: VFC<Omit<ContentsType, "onError">> = ({
+  onSubmit,
+}) => {
   const onError = (errors: any) => {
     console.log(`execute: onError`);
     console.log(errors);
